Update Homework19Tests.java
Changed Thread.sleep from 10000 to 2000 everywhere except after the playList was created. Before with 10000 all over it was giving the Fail Test, too much time. After I reset everywhere to 2000 I got another Fail due to the success messages being displayed at the same time and the actual value of the message was created playlist instead of deleted. By increasing the sleep time after creation of playlist to 10000 it allowed for the correct success message to be stored in the variable.